requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
neo4j==5.14.1
python-dotenv==1.0.0
requests==2.31.0
groq==0.3.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
python-dateutil==2.8.2
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.12.1
email-validator==2.1.0.post1
httpx==0.25.2